requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
black==21.8b0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.4
idna==3.2
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.2
MarkupSafe==1.1.1
matplotlib==3.4.3
multidict==5.1.0
mypy-extensions==0.4.3
networkx==2.6.2
numpy==1.21.2
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
pyee==7.0.4
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
PyWavelets==1.1.1
regex==2021.8.28
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
slackclient==2.9.3
slackeventsapi==2.2.1
tifffile==2021.8.30
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==1.0.1
yarl==1.6.3
